refactor: organize resources into dedicated registry pattern
Create a dedicated resources module that follows the same decorator-based registry pattern already used for tools and prompts. This change:
- Moves the Panther config resource to its own module
- Adds an mcp_resource decorator for consistent resource registration
- Centralizes resource management through a registry system
- Improves code organization and maintainability
- Sets the foundation for easier unit testing of resources
