+++ src/mcp_panther/__init__.py
@@ -1,0 +1,7 @@
+"""
+MCP Panther - An MCP server for interacting with Panther Security Platform.
+"""
+
+from .server import mcp, main, MCP_SERVER_NAME
+
+__all__ = ["mcp", "main", "MCP_SERVER_NAME"]
