+++ src/mcp_panther/panther_mcp_core/__init__.py
@@ -1,0 +1,14 @@
+"""
+Core functionality for the Panther MCP server.
+
+This package contains all the core functionality for the Panther MCP server,
+including API clients, tools, prompts, and resources.
+"""
+
+# Define all subpackages that should be available when importing this package
+__all__ = ["tools", "prompts", "resources"]
+
+# Ensure all subpackages are importable
+from . import tools
+from . import prompts
+from . import resources
