+++ src/mcp_panther/panther_mcp_core/tools/__init__.py
@@ -1,0 +1,17 @@
+"""
+Package for Panther MCP tools.
+
+This package contains all the tool functions available for Panther through MCP.
+All tool modules are imported here to ensure their decorators are processed.
+"""
+
+# Define all modules that should be available when importing this package
+__all__ = ["alerts", "rules", "data_lake", "sources", "metrics", "users"]
+
+# Import all tool modules to ensure decorators are processed
+from . import alerts
+from . import rules
+from . import data_lake
+from . import sources
+from . import metrics
+from . import users
